moved bcryptjs and jwt to helpers folder
