+++ controllers/userController.js
@@ -6,54 +6,54 @@ mongoose.connect('mongodb://localhost:27017/steam-clone-db', {
 const User = require('../models/user');
 
 class UserController {
-    static async getAll(req, res, next) {
+    static async getAll(request, response, next) {
         try {
             let users = await User.find({});
-            res.status(200).json({users});
+            response.status(200).json({users});
         } catch (error) {
             next({code:400, message: error.message});
         }
     }
 
-    static async getById(req, res, next) {
+    static async getById(request, response, next) {
         try {
-            let user = await User.findById(req.params.id);
-            res.status(200).json({user});
+            let user = await User.findById(request.params.id);
+            response.status(200).json({user});
         } catch (error) {
             next({code:400, message: error.message});
         }
     }
 
-    static async create(req, res, next) {
+    static async create(request, response, next) {
         try {
-            let user = await User.find({"email": req.body.email})
+            let user = await User.find({"email": request.body.email})
             if (user.length === 1) {
                 next({ code: 400, message: `This email is already registered!` });
             } else {
                 let newUser = {
-                    username: req.body.username,
-                    email: req.body.email,
-                    password: req.body.password
+                    username: request.body.username,
+                    email: request.body.email,
+                    password: request.body.password
                 }
                 let user = await User.create(newUser);
-                res.status(201).json({message: `User created successfully!`, user});
+                response.status(201).json({message: `User created successfully!`, user});
             }
         } catch (error) {
             next({code:400, message: error.message});
         }
     }
 
-    static async update(req, res, next) {
+    static async update(request, response, next) {
         try {
-            let user = await User.findById(req.params.id)
+            let user = await User.findById(request.params.id)
             if (user) {
                 let newUser = {
-                    username: req.body.username,
-                    email: req.body.email,
-                    password: req.body.password
+                    username: request.body.username,
+                    email: request.body.email,
+                    password: request.body.password
                 }
-                let user = await User.updateOne({_id: req.params.id}, newUser);
-                res.status(200).json({message: `User updated successfully!`, user});
+                let user = await User.updateOne({_id: request.params.id}, newUser);
+                response.status(200).json({message: `User updated successfully!`, user});
             } else {
                 next({code:404, message: `User not found!`});
             }
@@ -62,12 +62,12 @@ class UserController {
         }
     }
 
-    static async delete(req, res, next) {
+    static async delete(request, response, next) {
         try {
-            let user = await User.findById(req.params.id)
+            let user = await User.findById(request.params.id)
             if (user) {
-                let user = await User.deleteOne({_id: req.params.id})
-                res.status(200).json({message: `User deleted successfully!`, user});
+                let user = await User.deleteOne({_id: request.params.id})
+                response.status(200).json({message: `User deleted successfully!`, user});
             } else {
                 next({code:404, message: `User not found!`});
             }
