+++ helpers/bcrypt.js
